requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
Deprecated==1.2.13
distro==1.8.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
netaddr==0.8.0
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
python-dotenv==0.21.1
python-http-client==3.3.7
qasync==0.23.0
requests==2.28.2
sendgrid==6.9.7
shiboken2==5.15.2.1
shiboken6==6.4.2
starkbank-ecdsa==2.2.0
urllib3==1.26.14
wrapt==1.14.1
yarl==1.8.2
yookassa==2.3.4
